Update to Laravel 11
This is a jump from Laravel 4.2 to Laravel 11. I've tested
a lot of functionality, but there's likely more things that
need some adjustments.

I've based this on going through all upgrade guides, comparing it
against a clean installation and comparing it against the state
of the "intern" repo.

simplesamlphp is replaced by onelogin/php-saml similar as in
the "intern" repo.

Update to PHP 8.3.

Simplify the dev setup, dropping the docker-only dev workflow
to reduce complexity.

The database will be manually changed to use utf8mb4.